backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.117.1",
    "pydantic[email]>=2.11.9",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    "supabase>=2.3.0",
    "google-api-python-client>=2.100.0",
]

[dependency-groups]
dev = [
    "ruff>=0.6",
    "pytest>=8",
    "requests>=2.31.0",
]

[tool.pytest.ini_options]
addopts = "-q"

[tool.ruff]
line-length = 127
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "C90"]
fixable = ["ALL"]
